+++ firebase-dataconnect/src/main/kotlin/com/google/firebase/dataconnect/MutationRef.kt
@@ -36,7 +36,7 @@ import kotlinx.serialization.modules.SerializersModule
 public interface MutationRef<Data, Variables> : OperationRef<Data, Variables> {
   override suspend fun execute(): MutationResult<Data, Variables>
 
-  @FirebaseDataConnectExperimental
+  @ExperimentalFirebaseDataConnect
   override fun copy(
     operationName: String,
     variables: Variables,
@@ -47,14 +47,14 @@ public interface MutationRef<Data, Variables> : OperationRef<Data, Variables> {
     variablesSerializersModule: SerializersModule?,
   ): MutationRef<Data, Variables>
 
-  @FirebaseDataConnectExperimental
+  @ExperimentalFirebaseDataConnect
   override fun <NewVariables> withVariablesSerializer(
     variables: NewVariables,
     variablesSerializer: SerializationStrategy<NewVariables>,
     variablesSerializersModule: SerializersModule?,
   ): MutationRef<Data, NewVariables>
 
-  @FirebaseDataConnectExperimental
+  @ExperimentalFirebaseDataConnect
   override fun <NewData> withDataDeserializer(
     dataDeserializer: DeserializationStrategy<NewData>,
     dataSerializersModule: SerializersModule?,