+++ firebase-dataconnect/src/main/kotlin/com/google/firebase/dataconnect/core/MutationRefImpl.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-@file:OptIn(com.google.firebase.dataconnect.FirebaseDataConnectExperimental::class)
+@file:OptIn(com.google.firebase.dataconnect.ExperimentalFirebaseDataConnect::class)
 
 package com.google.firebase.dataconnect.core